+++ __tests__/nodes/makeData.js
@@ -1,0 +1,10 @@
+const timestamp = require('jest-mock-now')();
+const makeData = require('../../scripts/makeData');
+
+test('makeData to match snapshot for nodes', () => {
+  expect(makeData('../src/nodes')).toMatchSnapshot();
+});
+
+test('makeData to match snapshot for binary-operator-printers', () => {
+  expect(makeData('../src/binary-operator-printers')).toMatchSnapshot();
+});
