+++ src/binary-operator-printers/index.js
@@ -1,4 +1,4 @@
-/* This file was automatically generated on Tue Jul 02 2019 20:17:37 GMT+1200 (New Zealand Standard Time) */
+/* This file was automatically generated on 1562128164 */
 
 /* eslint-disable global-require */
